backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
lightgbm==4.5.0
pandas==2.2.3
pyarrow==19.0.1
numpy==2.2.1
scikit-learn==1.6.1
pydantic==2.10.4
python-multipart==0.0.20
sentry-sdk[fastapi]==2.19.2
ruff==0.8.6
pytest==8.3.4
httpx==0.28.1
PyJWT==2.9.0
stripe==11.4.1
